refactor(worker): move lobby read cleanup into service
